fix(services): Always provide report definitions for all plugins
Make sure that always report definitions for all existing reporter
plugins are present even if there is no admin configuration, or the
configuration does not contain a `reporter` section.
